core/serializers.py
from rest_framework import serializers
from .models import *

class BranchSerializer(serializers.ModelSerializer):
    class Meta:
        model = Branches
        fields = '__all__'

class WorkerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Workers
        fields = '__all__'

class CheksSerializer(serializers.ModelSerializer):
    class Meta:
        model = Checks
        fields = '__all__'
    
class DocsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Docs
        fields = '__all__'