requirements.txt
asgiref==3.7.2
boto3==1.28.49
botocore==1.31.49
Django==4.2.5
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-filter==23.2
django-storages==1.14
djangorestframework==3.14.0
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==1.26.16
